avfilter/estdif: avoid signed overflow in slice boundary calculation
deinterlace_slice() computed per-thread row boundaries with int
multiplication height * (jobnr + 1). With a tall frame and many filter
threads the product overflows signed int before the division by nb_jobs.

Use int64_t for the intermediate product before converting back to int
row indices.

Found-by: Kery (Qi Kery <qikeyu2001@outlook.com>)
(cherry picked from commit 200e0cba677b0d478576081ecf12af92b7a4a2d2)

+++ libavfilter/vf_estdif.c
@@ -356,8 +356,8 @@ static int deinterlace_slice(AVFilterContext *ctx, void *arg,
         const int height = s->planeheight[plane];
         const int src_linesize = in->linesize[plane];
         const int dst_linesize = out->linesize[plane];
-        const int start = (height * jobnr) / nb_jobs;
-        const int end = (height * (jobnr+1)) / nb_jobs;
+        const int start = (int)((int64_t)height * jobnr / nb_jobs);
+        const int end = (int)((int64_t)height * (jobnr + 1) / nb_jobs);
         const uint8_t *prev_line, *prev2_line, *next_line, *next2_line, *in_line;
         const uint8_t *prev3_line, *next3_line;
         uint8_t *out_line;
